UsersService/requirements.txt
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.6.2
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
multidict==6.1.0
pamqp==3.3.0
propcache==0.2.1
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
yarl==1.18.3
